Move `checked_binomial`
Just cut/paste, and fixed imports.
It is moved because it is now needed even without heap-allocation.
